Bug: 347831320
Merged-In: I008d7ed6df14c61b7e88dec95501e30de40b297f

+++ docs/examples/ImageInfo_minRowBytes.cpp
@@ -1,7 +1,6 @@
 // Copyright 2019 Google LLC.
 // Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 #include "tools/fiddle/examples.h"
-// HASH=897230ecfb36095486beca324fd369f9
 REG_FIDDLE(ImageInfo_minRowBytes, 256, 256, true, 0) {
 void draw(SkCanvas* canvas) {
     for (int shift = 24; shift < 31; ++shift) {
@@ -11,7 +10,7 @@ void draw(SkCanvas* canvas) {
         size_t minRowBytes = imageInfo.minRowBytes();
         bool widthTooLarge = !minRowBytes;
         SkDebugf("RGBA_F16 width %d (0x%08x) %s\n",
-                width, width, widthTooLarge ? "too large" : "OK");
+                 width, (unsigned int)width, widthTooLarge ? "too large" : "OK");
     }
 }
 }  // END FIDDLE
